+++ src/components/circuitDisplay/command.stories.js
@@ -38,7 +38,7 @@ const Template = (args) => ({
     }
   },
   template: `<div :class="[args.darkTheme ? 'theme-mode-dark' : 'theme-mode-light']">
-    <div class="circuit-display-container theme_variables">
+    <div class="circuit-display-container theme_variables border-box">
       <div style="justify-content: flex-start" class="circuit-container circuit-preview circuit_variables condensed">
         <div class="circuit-inner-scroll" data-cy="command-container">
           <circuit-command :ref="'commandRef'" v-bind="args" @mounted="onMounted">