+++ src/components/gateInfo/infoModal.vue
@@ -84,7 +84,7 @@ export default {
     bottom: 0;
     left: 0;
     right: 0;
-    z-index: 1;
+    z-index: 10;
 }
 .info-modal-backdrop{
     position: absolute;